requirements.txt
discord.py==2.5.2
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
httplib2==0.22.0